respa_admin: Fix tests
